Add "insert <unicode character name>" commands. (#9)
This adds commands for every character in the English Unicode CLDR thingy.  It reveals a bunch of issues with our autocomplete algorithm, but those can be dealt with in a future PR.

The Unicode CLDR data was retrieved from the [cldr-json][] project.  The file `src/cldr_annotations.rs` is generated by `build-cldr-annotations.js`.

[cldr-json]: https://github.com/unicode-org/cldr-json